Student Assistant Chatbot/requirements.txt
streamlit
smolagent
groq
google-generativeai
python-dotenv
